[prepare.rs] Move module-specific tests to module file
* Move module-specific tests to `prepare.rs` module file.

* Drop multiple `use` calls in tests and use a test-module level
`use super::*`.
